Simplify folder structure in tests
