--- tests/import/mod.rs
@@ -1,1 +1,0 @@
-pub mod user_service;
